fix(sessioncatalog): harden persisted projection repair
Problem:
The initial projection repair exceeded repository file-size and comment budgets, and its in-memory regression test did not cover a persisted catalog reopening after incomplete rows were stored.

Root cause:
The integrity query and regression test were added to files already at their capacity limits, while the test kept the same Catalog instance alive across each simulated projection loss.

Fix:
Move the ready-directory integrity check and regression coverage into focused files. Gate all indexed integrity scans on the matching ready signature, read the invariants from one SQLite snapshot, and reopen the disk-backed catalog before repairing missing session and topic rows.

Verification:
- go run ./tools/repolint
- go test ./internal/sessioncatalog/ -count=1
- go test -race ./internal/sessioncatalog/ -count=1
- go vet ./...
- go test ./... -count=1
- cd desktop && go test ./... -run 'Test(ListProjectTopics|ProjectTreeSnapshot|UpgradeLegacy|SessionCatalog|ReconcileRepairsPersistedReadyDirectory).*' -count=1
